dailyfresh/requirements.txt
alipay-sdk-python==3.3.398
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
Authlib==1.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0
django-haystack==3.1.1
django-redis==5.2.0
django-tinymce==3.4.0
dnspython==2.2.1
eventlet==0.33.0
fdfs-client-py==1.2.6
greenlet==1.1.2
idna==3.3
jieba==0.42.1
kombu==5.2.4
mutagen==1.45.1
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pycryptosat @ file:///D:/pycryptosat-0.6.1.1-cp38-cp38-win_amd64.whl
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
redis==4.3.1
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
wincertstore==0.2
wrapt==1.14.1
